remove audio init | clearer up logs | clean up config
