+++ lib/scripts/core/settings.ts
@@ -5,7 +5,7 @@ export const MAX_PIXEL_COUNT = 2560 * 1440;
 const FREE_BLOCKS_COUNT = 12;
 const AUTO_RESTART = true;
 const AUTO_START = false;
-const SHOW_BLOCK = true;
+const SHOW_BLOCK = false;
 const DEFAULT_POSITION = [-20, 18, 20];
 const DEFAULT_LOOKAT_POSITION = [0, 0, 0];
 export const WEBGL_OPTS = { antialias: true, alpha: false, powerPreference: 'low-power' as WebGLPowerPreference };
