+++ src/index.ts
@@ -1,5 +1,6 @@
 import GUI from 'lil-gui';
 
+// import { loadTowerAnimation, removeTowerAnimation, setAnimationProperties, setAnimationState } from '../lib';
 import { loadTowerAnimation, removeTowerAnimation, setAnimationProperties, setAnimationState } from '@tari-project/tari-tower';
 
 if (import.meta.env.MODE === 'development') {
