+++ tsconfig.app.json
@@ -24,5 +24,6 @@
 		"types": ["vite/client", "node"]
 	},
 	"include": ["src"],
+	"exclude": ["node_modules", "dist"],
 	"references": [{ "path": "./tsconfig.node.json" }]
 }
