+++ vite.config.ts
@@ -1,4 +1,5 @@
 import { defineConfig } from 'vite';
+import { resolve } from 'path';
 import eslint from '@nabla/vite-plugin-eslint';
 import tsconfigPaths from 'vite-tsconfig-paths';
 import dts from 'vite-plugin-dts';
@@ -7,26 +8,24 @@ export default defineConfig({
 	server: {
 		port: 3001,
 	},
-	plugins: [dts({ include: ['lib'], rollupTypes: true, tsconfigPath: './tsconfig.app.json' }), tsconfigPaths(), eslint({ eslintOptions: { cache: false, fix: true } })],
+	plugins: [
+		tsconfigPaths(),
+		eslint({ eslintOptions: { cache: false, fix: true } }),
+		dts({ include: ['lib'], exclude: ['src'], rollupTypes: true, tsconfigPath: resolve(__dirname, 'tsconfig.app.json') }),
+	],
 	build: {
+		emptyOutDir: true,
 		lib: {
-			entry: 'lib/index.ts',
+			entry: resolve(__dirname, 'lib/index.ts'),
 			name: '@tari-project/tari-tower',
 			formats: ['es'],
 		},
 		rollupOptions: {
-			input: 'lib/index.ts',
-			external: ['three', 'min-signal'],
+			input: resolve(__dirname, 'lib/index.ts'),
 			output: {
 				entryFileNames: '[name].js',
 				assetFileNames: 'assets/[name][extname]',
-				globals: {
-					'three': 'Three',
-					'min-signal': 'MinSignal',
-				},
 			},
-			makeAbsoluteExternalsRelative: true,
-			preserveEntrySignatures: 'exports-only',
 		},
 	},
 });
